user.go: simplify the check for main module
Update the newUserLocator function to simply call
l.locate(info.Main.Path) and in case of error return the null locator.

In the error message, replace "application" with "main module" and use
info.Main instead of AppName() so that it is consistent with the other
locators.

Remove the now unused isInUserData function.